# docs/lyc-api-openapi.yaml
openapi: 3.0.3
info:
  title: LYC 小宇宙学堂 API
  description: LYC 儿童启蒙学习平台后端 API，涵盖认证、内容管理、媒体播放、AI 分析、学习进度等模块。
  version: 1.0.0
  contact:
    name: LYC Team
servers:
  - url: http://127.0.0.1:3002
    description: 本地开发
  - url: https://api.lyc.example.com
    description: 生产环境
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 小程序端登录后获取的 JWT Token
    adminKey:
      type: apiKey
      in: header
      name: X-Admin-Key
      description: 管理员 API Key
  schemas:
...
    Category:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
          example: "英语跟读"
        subtitle:
          type: string
          nullable: true
        badge:
          type: string
          nullable: true
        icon:
          type: string
          nullable: true
        heroTitle:
          type: string
          nullable: true
        heroText:
          type: string
          nullable: true
        tone:
          type: string
          enum: [sky, sun, moon, mint]
        sortOrder:
          type: integer
        status:
          type: string
          enum: [active, deleted]
        createdAt:
          type: string
          format: date-time
    Course:
      type: object
      properties:
        id:
          type: string
        categoryId:
          type: string
        title:
          type: string
          example: "春晓"
        subtitle:
          type: string
          nullable: true
        summary:
          type: string
          nullable: true
        durationText:
          type: string
          nullable: true
          example: "约 6 分钟"
        ageRange:
          type: string
          nullable: true
          example: "5-8 岁"
        difficulty:
          type: string
          nullable: true
        reward:
          type: integer
        coverLabel:
          type: string
          nullable: true
        mentor:
          type: string
          nullable: true
        coverUrl:
          type: string
          nullable: true
        audioUrl:
          type: string
          nullable: true
        videoUrl:
          type: string
          nullable: true
        totalDurationMs:
          type: integer
        fullText:
          type: string
          nullable: true
        asrStatus:
          type: string
          enum: [none, processing, done, failed]
        segmentCount:
          type: integer
        status:
          type: string
          enum: [active, deleted]
        sortOrder:
          type: integer
        createdAt:
          type: string
          format: date-time
    Segment:
      type: object
      properties:
        id:
          type: string
        courseId:
          type: string
        segmentIndex:
          type: integer
        title:
          type: string
          nullable: true
        cue:
          type: string
          nullable: true
        text:
          type: string
        translation:
          type: string
          nullable: true
        tip:
          type: string
          nullable: true
        startTimeMs:
          type: integer
        endTimeMs:
          type: integer
        audioUrl:
          type: string
          nullable: true
        durationMs:
          type: integer
        sortOrder:
          type: integer
...
paths:
...
  /api/v1/content/categories:
    get:
      tags: [内容管理]
      summary: 获取分类列表
      responses:
        "200":
          description: 成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/Category"
    post:
      tags: [内容管理]
      summary: 创建分类
      security:
        - adminKey: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name:
                  type: string
                subtitle:
                  type: string
                badge:
                  type: string
                icon:
                  type: string
                heroTitle:
                  type: string
                heroText:
                  type: string
                tone:
                  type: string
                sortOrder:
                  type: integer
      responses:
        "200":
          description: 创建成功
  /api/v1/content/categories/{id}:
    put:
      tags: [内容管理]
      summary: 更新分类
      security:
        - adminKey: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                subtitle:
                  type: string
                tone:
                  type: string
                sortOrder:
                  type: integer
                status:
                  type: string
      responses:
        "200":
          description: 更新成功
    delete:
      tags: [内容管理]
      summary: 删除分类
      security:
        - adminKey: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: 删除成功
  /api/v1/content/courses:
    get:
      tags: [内容管理]
      summary: 获取课程列表
      parameters:
        - name: categoryId
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: pageSize
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: 成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  total:
                    type: integer
                  page:
                    type: integer
                  pageSize:
                    type: integer
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/Course"
    post:
      tags: [内容管理]
      summary: 创建课程
      security:
        - adminKey: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [title, categoryId]
              properties:
                categoryId:
                  type: string
                title:
                  type: string
                subtitle:
                  type: string
                summary:
                  type: string
                durationText:
                  type: string
                ageRange:
                  type: string
                difficulty:
                  type: string
                reward:
                  type: integer
                mentor:
                  type: string
                audioUrl:
                  type: string
                videoUrl:
                  type: string
                sortOrder:
                  type: integer
      responses:
        "200":
          description: 创建成功
  /api/v1/content/courses/{id}:
    get:
      tags: [内容管理]
      summary: 获取课程详情
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: 成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  course:
                    $ref: "#/components/schemas/Course"
                  segments:
                    type: array
                    items:
                      $ref: "#/components/schemas/Segment"
    put:
      tags: [内容管理]
      summary: 更新课程
      security:
        - adminKey: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                subtitle:
                  type: string
                summary:
                  type: string
                audioUrl:
                  type: string
                videoUrl:
                  type: string
                asrStatus:
                  type: string
                status:
                  type: string
                sortOrder:
                  type: integer
      responses:
        "200":
          description: 更新成功
    delete:
      tags: [内容管理]
      summary: 删除课程
      security:
        - adminKey: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: 删除成功
  /api/v1/content/courses/{courseId}/segments:
    post:
      tags: [内容管理]
      summary: 添加段落
      security:
        - adminKey: []
      parameters:
        - name: courseId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                segmentIndex:
                  type: integer
                title:
                  type: string
                cue:
                  type: string
                text:
                  type: string
                translation:
                  type: string
                tip:
                  type: string
                startTimeMs:
                  type: integer
                endTimeMs:
                  type: integer
                sortOrder:
                  type: integer
      responses:
        "200":
          description: 创建成功
  /api/v1/content/segments/{segmentId}:
    put:
      tags: [内容管理]
      summary: 更新段落
      security:
        - adminKey: []
      parameters:
        - name: segmentId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                text:
                  type: string
                translation:
                  type: string
                tip:
                  type: string
                startTimeMs:
                  type: integer
                endTimeMs:
                  type: integer
      responses:
        "200":
          description: 更新成功
    delete:
      tags: [内容管理]
      summary: 删除段落
      security:
        - adminKey: []
      parameters:
        - name: segmentId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: 删除成功

  # ========== 文件上传 ==========
...
  /api/v1/media/stream/{key}:
    get:
      tags: [媒体播放]
      summary: 流式播放
      description: 音视频流式播放，支持 Range 请求（HTTP 206），可拖动进度条。
      parameters:
        - name: key
          in: path
          required: true
          schema:
            type: string
        - name: Range
          in: header
          description: "字节范围，如 bytes=0-1048575"
          schema:
            type: string
      responses:
        "200":
          description: 完整文件
          content:
            audio/mpeg:
              schema:
                type: string
                format: binary
            audio/mp4:
              schema:
                type: string
                format: binary
            video/mp4:
              schema:
                type: string
                format: binary
        "206":
          description: 部分内容（Range 请求）
          headers:
            Content-Range:
              schema:
                type: string
                example: "bytes 0-1048575/10485760"
            Accept-Ranges:
              schema:
                type: string
                example: bytes
  /api/v1/media/signed-url/{fileId}:
    get:
      tags: [媒体播放]
      summary: 获取签名 URL
      parameters:
        - name: fileId
          in: path
          required: true
          schema:
            type: string
        - name: expires
          in: query
          schema:
            type: integer
            default: 3600
      responses:
        "200":
          description: 成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  url:
                    type: string
                  expires:
                    type: integer
  /api/v1/media/info/{fileId}:
    get:
      tags: [媒体播放]
      summary: 获取文件信息
      parameters:
        - name: fileId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: 成功

  # ========== AI 分析 ==========
...